Refactored app and chain stores away to itw own files
