+++ src/pages/home.tsx
@@ -2,7 +2,7 @@ import { ChainProvider, ReactiveDotProvider } from '@reactive-dot/react';
 import { useProxy } from 'valtio/utils';
 import { ConfigProvider, useConfig } from '~/api/config2';
 import { IdentityRegistrarComponent } from '~/components/identity-registrar';
-import { chainStore as _chainStore } from "~/store"
+import { chainStore as _chainStore } from '~/store/ChainStore';
 
 function Home() {
   return (
